requirements.txt
django
pillow
django-crispy-forms
django-pagedown
django-taggit
python-dotenv
mysqlclient
